devtools/shared/webconsole/test/chrome/test_network_post.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for the network actor (POST request)</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="common.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for the network actor (POST request)</p>

<iframe src="http://example.com/chrome/devtools/shared/webconsole/test/chrome/network_requests_iframe.html"></iframe>

<script class="testbody" type="text/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();

async function startTest()
{
  removeEventListener("load", startTest);

  const {state} = await attachConsoleToTab(["NetworkActivity"]);
  onAttach(state);
}

function onAttach(aState)
{
  info("test network POST request");

  onNetworkEvent = onNetworkEvent.bind(null, aState);
  aState.webConsoleFront.on("serverNetworkEvent", onNetworkEvent);
  onNetworkEventUpdate = onNetworkEventUpdate.bind(null, aState);
  aState.dbgClient.on("networkEventUpdate", onNetworkEventUpdate);

  const iframe = document.querySelector("iframe").contentWindow;
  iframe.wrappedJSObject.testXhrPost();
}

function onNetworkEvent(aState, aPacket)
{
  info("checking the network event packet");

  const netActor = aPacket.eventActor;

  checkObject(netActor, {
    actor: /[a-z]/,
    startedDateTime: /^\d+\-\d+\-\d+T.+$/,
    url: /data\.json/,
    method: "POST",
  });

  aState.netActor = netActor.actor;

  aState.webConsoleFront.off("serverNetworkEvent", onNetworkEvent);
}

const updates = [];

function onNetworkEventUpdate(aState, aPacket)
{
  info("received networkEventUpdate " + aPacket.updateType);
  is(aPacket.from, aState.netActor, "networkEventUpdate actor");

  updates.push(aPacket.updateType);

  let expectedPacket = null;

  switch (aPacket.updateType) {
    case "requestHeaders":
    case "responseHeaders":
      ok(aPacket.headers > 0, "headers > 0");
      ok(aPacket.headersSize > 0, "headersSize > 0");
      break;
    case "requestCookies":
      expectedPacket = {
        cookies: 3,
      };
      break;
    case "requestPostData":
      ok(aPacket.dataSize > 0, "dataSize > 0");
      break;
    case "responseStart":
      expectedPacket = {
        response: {
          httpVersion: /^HTTP\/\d\.\d$/,
          status: "200",
          statusText: "OK",
          headersSize: /^\d+$/,
        },
      };
      break;
    case "securityInfo":
      expectedPacket = {
        state: "insecure",
      };
      break;
    case "responseCookies":
      expectedPacket = {
        cookies: 0,
      };
      break;
    case "responseContent":
      expectedPacket = {
        mimeType: "application/json",
        contentSize: /^\d+$/,
      };
      break;
    case "eventTimings":
      expectedPacket = {
        totalTime: /^\d+$/,
      };
      break;
    default:
      ok(false, "unknown network event update type: " +
         aPacket.updateType);
      return;
  }

  if (expectedPacket) {
    info("checking the packet content");
    checkObject(aPacket, expectedPacket);
  }

  if (updates.indexOf("responseContent") > -1 &&
      updates.indexOf("eventTimings") > -1) {
    aState.dbgClient.off("networkEventUpdate",
                                    onNetworkEvent);

    onRequestHeaders = onRequestHeaders.bind(null, aState);
    aState.webConsoleFront.getRequestHeaders(aState.netActor,
                                    onRequestHeaders);
  }
}

function onRequestHeaders(aState, aResponse)
{
  info("checking request headers");

  ok(aResponse.headers.length > 0, "request headers > 0");
  ok(aResponse.headersSize > 0, "request headersSize > 0");
  ok(!!aResponse.rawHeaders.length, "request rawHeaders available");

  checkHeadersOrCookies(aResponse.headers, {
    Referer: /network_requests_iframe\.html/,
    Cookie: /bug768096/,
  });

  checkRawHeaders(aResponse.rawHeaders, {
    Referer: /network_requests_iframe\.html/,
    Cookie: /bug768096/,
  });

  onRequestCookies = onRequestCookies.bind(null, aState);
  aState.webConsoleFront.getRequestCookies(aState.netActor,
                                  onRequestCookies);
}

function onRequestCookies(aState, aResponse)
{
  info("checking request cookies");

  is(aResponse.cookies.length, 3, "request cookies length");

  checkHeadersOrCookies(aResponse.cookies, {
    foobar: "fooval",
    omgfoo: "bug768096",
    badcookie: "bug826798=st3fan",
  });

  onRequestPostData = onRequestPostData.bind(null, aState);
  aState.webConsoleFront.getRequestPostData(aState.netActor,
                                   onRequestPostData);
}

function onRequestPostData(aState, aResponse)
{
  info("checking request POST data");

  checkObject(aResponse, {
    postData: {
      text: /^Hello world! foobaz barr.+foobaz barr$/,
    },
    postDataDiscarded: false,
  });

  is(aResponse.postData.text.length, 552, "postData text length");

  onResponseHeaders = onResponseHeaders.bind(null, aState);
  aState.webConsoleFront.getResponseHeaders(aState.netActor,
                                   onResponseHeaders);
}

function onResponseHeaders(aState, aResponse)
{
  info("checking response headers");

  ok(aResponse.headers.length > 0, "response headers > 0");
  ok(aResponse.headersSize > 0, "response headersSize > 0");
  ok(!!aResponse.rawHeaders, "response rawHeaders available");

  checkHeadersOrCookies(aResponse.headers, {
    "content-type": /^application\/(json|octet-stream)$/,
    "content-length": /^\d+$/,
  });

  checkRawHeaders(aResponse.rawHeaders, {
    "content-type": /^application\/(json|octet-stream)$/,
    "content-length": /^\d+$/,
  });

  onResponseCookies = onResponseCookies.bind(null, aState);
  aState.webConsoleFront.getResponseCookies(aState.netActor,
                                  onResponseCookies);
}

function onResponseCookies(aState, aResponse)
{
  info("checking response cookies");

  is(aResponse.cookies.length, 0, "response cookies length");

  onResponseContent = onResponseContent.bind(null, aState);
  aState.webConsoleFront.getResponseContent(aState.netActor,
                                   onResponseContent);
}

function onResponseContent(aState, aResponse)
{
  info("checking response content");

  checkObject(aResponse, {
    content: {
      text: /"test JSON data"/,
    },
    contentDiscarded: false,
  });

  onEventTimings = onEventTimings.bind(null, aState);
  aState.webConsoleFront.getEventTimings(aState.netActor,
                                onEventTimings);
}

function onEventTimings(aState, aResponse)
{
  info("checking event timings");

  checkObject(aResponse, {
    timings: {
      blocked: /^-1|\d+$/,
      dns: /^-1|\d+$/,
      connect: /^-1|\d+$/,
      send: /^-1|\d+$/,
      wait: /^-1|\d+$/,
      receive: /^-1|\d+$/,
    },
    totalTime: /^\d+$/,
  });

  closeDebugger(aState, function() {
    SimpleTest.finish();
  });
}

addEventListener("load", startTest);
</script>
</body>
</html>
